v3.5: systematic logic error audit and fixes
CRITICAL:
- IMU.c: if_get_offset made static (yaw was frozen after first calibration)
- IMU.c: replace dead nowtime with DWT->CYCCNT timing (AHRS quaternion frozen)
- app_sensor.c: remove redundant Car_Attitude_Update_Input (stale encoder data overwrite)
- car_control.c: epsilon-based float zero comparison (mode selection failure)
- car_attitude.c: epsilon-based stop detection (micro-motion bug)
- car_control.c: clamp asin domain [-1,1] (NaN on TO_POINT)

MEDIUM:
- IMU.c: fix strict aliasing UB in invSqrt1 (union instead of ptr cast)
- IMU.c: use || instead of && for error correction trigger
- Motor.c: remove pointless int->float roundtrip in Set_V_Real
- freertos.c: remove duplicate FreeRTOS.h include

Documentation: Phase 8 audit table in README

+++ Core/Src/Motor.c
@@ -107,8 +107,8 @@ Set_PID(&motor_RightRear.v_pid, P_RR, I_RR, D_RR);
  * @param v_real 目标速度mm/s,float
  */
 void Motor_Set_V_Real(motor* motor,float v_real){
-    int v_enc=V_REAL_TO_ENC*v_real;
-    Motor_Set_V_Enc(motor,v_enc);
+    float v_enc = V_REAL_TO_ENC * v_real;
+    Motor_Set_V_Enc(motor, v_enc);
 }
 
 /*!
